requirements.txt
transformers==4.35.2
torch==2.0.1
torchvision==0.15.2
gradio==4.19.0
peft==0.7.1
accelerate==0.25.0
pillow==10.0.0
numpy==1.24.3
safetensors==0.4.1